requirements/test.txt
-r prod.txt
django-debug-toolbar==4.4.6
flake8-commas==4.0.0
flake8-import-order==0.18.2
flake8-quotes==3.4.0
parameterized==0.9.0
pep8-naming==0.14.1